Added support for quality background colors
