+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/Passport/NFTIcons/NFTIconComponentModel.cs
@@ -8,5 +8,6 @@ public class NFTIconComponentModel : BaseComponentModel
     public string type;
     public string marketplaceURI;
     public string name;
+    public string rarity;
     public string imageURI;
 }
